services/media-service/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
Pillow
supabase
drf-yasg
python-dotenv
psycopg2-binary
django-redis
redis
gunicorn
confluent-kafka

